requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.2.17
google-auth==2.35.0
google-auth-oauthlib==1.2.1
idna==3.9
psycopg2-binary
django-environ
imageio==2.35.1
imageio-ffmpeg==0.5.1
MarkupSafe==2.1.5
moviepy==1.0.3
numpy
oauthlib==3.2.2
boto3
django-storages
opencv-python==4.10.0.84
pandas>=2.0.0
Pillow==9.5.0
proglog==0.1.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
stripe==10.11.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
